MYR-398 Replace the activity notifier's WaitGroup with a mutex-guarded drain
Wait ran on a different goroutine from the one starting workers — the bus's
delivery goroutine calls handleStatusChanged, and nothing orders that against
a shutdown calling Wait — so a WaitGroup's rule that an Add off zero must
happen-before Wait could not hold. The counter may then be read empty while an
accepted event is still on its way in, which is a drain that ends before the
last push. -race caught it on main after #363 and skipped the deploy.

Counting inflight workers under the mutex the waiter blocks on closes the
window, and stays reusable as a repeated quiesce point.
